requirements.txt
absl-py==1.3.0
asgiref==3.5.2
asttokens==2.2.1
astunparse==1.6.3
attrs==22.1.0
backcall==0.2.0
backports.zoneinfo==0.2.1
bel-resources==0.0.3
biopython==1.80
bioregistry==0.6.33
bitarray==2.6.2
build==0.9.0
cachetools==5.2.1
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
click-plugins==1.1.1
colorama==0.4.6
crispy-bootstrap5==0.7
curies==0.4.0
decorator==5.1.1
dill==0.3.6
dj-database-url==1.2.0
Django==4.1.3
django-crispy-forms==1.14.0
django-rest-framework==0.1.0
djangorestframework==3.14.0
docutils==0.17.1
executing==1.2.0
fastcore==1.5.27
flatbuffers
gast==0.4.0
google-auth==2.15.0
google-auth-oauthlib
google-pasta==0.2.0
greenlet==2.0.1
grpcio==1.51.1
gunicorn==20.1.0
h5py==3.7.0
humanize==4.4.0
idna==3.4
importlib-metadata==6.0.0
importlib-resources==5.10.1
ipython==8.7.0
jedi==0.18.2
joblib==1.2.0
jsonschema==4.17.3
keras
kora==0.9.20
libclang==15.0.6.1
Markdown==3.4.1
MarkupSafe==2.1.1
matplotlib-inline==0.1.6
mmcif-pdbx==2.0.1
more-click==0.1.2
more-itertools==9.0.0
multiprocess==0.70.14
multisplitby==0.0.1
networkx==2.8.8
numpy==1.23.5
oauthlib==3.2.2
openbabel==3.1.1.1
opt-einsum==3.3.0
packaging==21.3
pandarallel==1.6.3
pandas==1.5.2
parso==0.8.3
pathos==0.3.0
pdb2pqr==3.5.2
Pebble==5.0.3
pep517==0.13.0
pickleshare==0.7.5
Pillow==9.3.0
pkgutil-resolve-name==1.3.10
pox==0.3.2
ppft==1.7.6.6
prompt-toolkit==3.0.33
propka==3.4.0
protobuf
psutil==5.9.4
psycopg2-binary==2.9.5
pure-eval==0.2.2
py3Dmol==1.8.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pybel==0.15.5
pydantic==1.10.2
Pygments==2.13.0
pyparsing==3.0.9
pypi==2.1
pyrsistent==0.19.2
pystow==0.4.7
python-dateutil==2.8.2
PyTrie==0.4.0
pytz==2022.6
ratelimit==2.2.1
rdkit==2022.9.2
requests==2.28.1
requests-file==1.5.1
requests-oauthlib==1.3.1
rsa==4.9
scikit-learn==1.2.0
scipy==1.10.0
six==1.16.0
scikit-learn
sortedcontainers==2.4.0
SQLAlchemy==1.4.44
sqlparse==0.4.3
stack-data==0.6.2
tabulate==0.9.0
tensorboard
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.14.0
tensorflow-estimator==2.14.0
tensorflow-io-gcs-filesystem==0.29.0
termcolor==2.2.0
threadpoolctl==3.1.0
tomli==2.0.1
tqdm==4.64.1
traitlets==5.6.0
typing-extensions==4.4.0
tzdata==2022.6
urllib3==1.26.13
wcwidth==0.2.5
Werkzeug==2.2.2
wrapt==1.14.1
zipp==3.11.0
